requirements.txt
numpy>=1.26,<2
pandas>=2.2,<3
# numba  # Se elimina temporalmente por incompatibilidad con Python 3.13
matplotlib>=3.8,<4
scikit-learn>=1.3.0
PyYAML>=6.0
